requirements.txt
# CertLens — AI Skill Inference Core
# Python 3.10+

# Google Gemini API (vision + text)
google-generativeai==0.8.3

# PDF to image conversion
pymupdf==1.25.3

# Image processing
Pillow==11.1.0

# Environment variable management
python-dotenv==1.0.1

reportlab==4.2.5

# Web UI
streamlit>=1.45.0